refactor: 팀원 추가 컴포넌트 리펙토링
