+++ idearly/src/components/AddTeamMembers/AddTeamMembers.styles.tsx
@@ -1,23 +1,5 @@
-import { HStack, Tag, TagLabel } from "@chakra-ui/react";
+import { Tag } from "@chakra-ui/react";
 import styled from "@emotion/styled";
-
-export const HStackWrapper = styled(HStack)`
-  margin-bottom: 0.7rem;
-`;
-
-export const TagWrapper = styled(Tag)`
-  background-color: #CBD5E0;
-  color: #4A5568;
-`;
-
-export const TagLabelName = styled(TagLabel)`
-  font-size: 0.8rem;
-`;
-
-
-export const TagLabeEmail = styled(TagLabel)`
-  font-size: 0.6rem;
-`;
 
 export const ShowUserTag = styled(Tag)`
   width: 100%;
