+++ idearly/src/pages/TeamMatchingPage/TeamMatchingPage.styles.tsx
@@ -1,5 +1,5 @@
 import styled from "@emotion/styled";
-import { Card, CardBody, CardFooter, CardHeader, Heading, Text } from "@chakra-ui/react";
+import { Card, CardBody, CardFooter, CardHeader, HStack, Heading, Tag, TagLabel, Text } from "@chakra-ui/react";
 
 export const TeamMathingWrapper = styled.div`
   width: 100%;
@@ -19,15 +19,30 @@ export const CardHeaderSection = styled(CardHeader)`
   text-align: center;
 `;
 
-export const TeamMatchingTitle = styled(Heading)`
-`;
+export const TeamMatchingTitle = styled(Heading)``;
 
-export const CardBodySection = styled(CardBody)`
-`;
+export const CardBodySection = styled(CardBody)``;
 
 export const CardFooterSection = styled(CardFooter)`
   display: flex;
   justify-content: center;
+`;
+
+export const HStackWrapper = styled(HStack)`
+  margin-bottom: 0.7rem;
+`;
+
+export const TagLabelName = styled(TagLabel)`
+  font-size: 0.8rem;
+`;
+
+export const TagLabeEmail = styled(TagLabel)`
+  font-size: 0.6rem;
+`;
+
+export const TagWrapper = styled(Tag)`
+  background-color: #CBD5E0;
+  color: #4A5568;
 `;
 
 export const MiniTitle = styled(Text)`
